requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pyyaml==6.0.1
django-filter==24.2